+++ Library/SkinInstaller.cpp
@@ -251,7 +251,7 @@ bool CopyFiles(const std::wstring& strFrom, const std::wstring& strTo, bool bMov
 	SHFILEOPSTRUCT fo =
 	{
 		nullptr,
-		bMove ? FO_MOVE : FO_COPY,
+		(UINT)(bMove ? FO_MOVE : FO_COPY),
 		tmpFrom.c_str(),
 		tmpTo.c_str(),
 		FOF_NO_UI | FOF_NOCONFIRMATION | FOF_ALLOWUNDO
